run_wps_process moved to common.py
