+++ tests/common.py
@@ -1,6 +1,6 @@
-from pywps import get_ElementMakerForVersion
+from pywps import get_ElementMakerForVersion, Service
 from pywps.app.basic import get_xpath_ns
-from pywps.tests import WpsClient, WpsTestResponse
+from pywps.tests import WpsClient, WpsTestResponse, assert_response_success
 
 VERSION = "1.0.0"
 WPS, OWS = get_ElementMakerForVersion(VERSION)
@@ -41,3 +41,16 @@ def get_output(doc):
             output[identifier_el.text] = data_el[0].text
 
     return output
+
+def run_wps_process(process, params):
+    client = client_for(Service(processes=[process]))
+    datainputs = params
+    resp = client.get(
+        service="wps",
+        request="Execute",
+        version="1.0.0",
+        identifier=process.identifier,
+        datainputs=datainputs,
+    )
+
+    assert_response_success(resp)
